control-plane/agents/shield-sword/requirements.txt
fastapi
uvicorn
httpx
pydantic
anthropic
pyyaml
